+++ oap-server/server-query-plugin/traceql-plugin/src/main/java/org/apache/skywalking/oap/query/traceql/rt/TraceQLQueryParams.java
@@ -33,6 +33,11 @@ public class TraceQLQueryParams {
     private String serviceName;
 
     /**
+     * Service instance name filter
+     */
+    private String serviceInstance;
+
+    /**
      * Span name filter
      */
     private String spanName;